feat: like service (incluindo teste springboot)

+++ src/main/java/tracklistd/api/Dto/Like/LikeResponseDTO.java
@@ -1,7 +1,7 @@
 package tracklistd.api.Dto.Like;
 
 public record LikeResponseDTO(
-                boolean liked,
-                Integer likesCount) {
+        boolean liked,
+        Long likesCount) {
 
 }
